split out random generators from base_fixture
