+++ cpp/tests/bitmask/bitmask_tests.cpp
@@ -25,6 +25,7 @@
 #include <cudf_test/column_utilities.hpp>
 #include <cudf_test/column_wrapper.hpp>
 #include <cudf_test/cudf_gtest.hpp>
+#include <cudf_test/random.hpp>
 #include <cudf_test/testing_main.hpp>
 
 #include <rmm/cuda_stream_view.hpp>
