+++ cpp/tests/iterator/pair_iterator_test_numeric.cu
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2023, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2024, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,8 @@
  * the License.
  */
 #include <tests/iterator/pair_iterator_test.cuh>
+
+#include <cudf_test/random.hpp>
 
 #include <rmm/exec_policy.hpp>
