+++ cpp/tests/iterator/value_iterator_test_transform.cu
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2023, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2024, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,8 @@
  * the License.
  */
 #include <tests/iterator/iterator_tests.cuh>
+
+#include <cudf_test/random.hpp>
 
 #include <thrust/functional.h>
 #include <thrust/host_vector.h>
